Avoid re-exporting files when content hasn't changed (#694)
Close 693.

+++ lib/i18n-js.rb
@@ -74,6 +74,12 @@ module I18nJS
     digest = Digest::MD5.hexdigest(contents)
     file_path = file_path.gsub(/:digest/, digest)
 
+    # Don't rewrite the file if it already exists and has the same content.
+    # It helps the asset pipeline or webpack understand that file wasn't changed.
+    if File.exist?(file_path) && File.read(file_path) == contents
+      return file_path
+    end
+
     File.open(file_path, "w") do |file|
       file << contents
     end
